requirements.txt
praw
feedparser
supabase
python-dotenv
httpx
anthropic